roles, cli, menu roles

+++ prj-website/src/jvmMain/kotlin/database/schema/ac_sessions.kt
@@ -4,7 +4,7 @@ const val session_id_length = 30
 
 object ac_sessions : Table(), CreatedAt, UpdatedAt {
     val id = varchar("id", session_id_length)
-    val user_id = long("user_id")
+    val user_id = integer("user_id")
     override val created_at = createdAt()
     override val updated_at = updatedAt()
     override val primaryKey = PrimaryKey(id)
